Rewrite plugin for Saleor 3.14
Over the past year several changes were made to Saleor, including:
- dropping the vatlayer plugin
- changing the behavior of plugin methods

These changes require a rewrite of most of the plugin, which
hopefully still retains the same functionality. For the plugin to
work correctly, it's necessary to use the "TAX_APP" strategy in the
tax setting of Saleor.

Using TAX_APP delegates tax calculation to our plugin, however,
unless the user has a valid VATIN, we actually use the
update_checkout_prices_with_flat_rates method from the FLAT_RATES
taxCalculationStrategy. This way we can:

- reverse charge VAT for EU businesses
- charge VAT for everyone else, according to their country

Special thanks to Sergey for the support while going through this.